+++ networkx/generators/degree_seq.py
@@ -428,7 +428,7 @@ def expected_degree_graph(w, seed=None, selfloops=True):
         while v < n and p > 0:
             if p != 1:
                 r = seed.random()
-                v += int(math.floor(math.log(r, 1 - p)))
+                v += math.floor(math.log(r, 1 - p))
             if v < n:
                 q = min(seq[v] * factor, 1)
                 if seed.random() < q / p:
